backend/core/requirements.txt
fastapi
uvicorn
requests
qdrant_client
sentence_transformers
langchain_text_splitters
pydantic
openai==0.28
python-multipart 
python-dotenv   
opencv-python-headless
opencv-python
minio
psycopg2-binary
bcrypt
motor 
python-jose[cryptography]
pdf2image
google-generativeai
google-cloud-vision
tensorflow==2.18.0
tf-keras